Add performance.now benchmarks

+++ src/test.c
@@ -4,7 +4,7 @@ char *test() {
 }
 
 float mathAverage(int *array, int size) {
-  float sum = 0;
+  int sum = 0;
 
   for (int i = 0; i < size; i++) {
     sum += array[i];
